+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 #### Improvements
 
-- Added support for `ln` in `snowflake.snowpark.functions`
+- Improved documentation for `Session.write_pandas` by making `use_logical_type` option more explicit.
 - Added support for specifying the following to `DataFrameWriter.save_as_table`:
   - `enable_schema_evolution`
   - `data_retention_time`
@@ -81,6 +81,7 @@
 - Added support for `DataFrameGroupBy.value_counts` and `SeriesGroupBy.value_counts`.
 - Added support for `Series.is_monotonic_increasing` and `Series.is_monotonic_decreasing`.
 - Added support for `Index.is_monotonic_increasing` and `Index.is_monotonic_decreasing`.
+- Added support for `pd.crosstab`.
 
 #### Improvements